Refactor : 리뷰 평점 소수점도 가능하게 수정 & 여행지 관심 상품에 담겼는지 정보 변경

+++ src/main/resources/schema.sql
@@ -105,7 +105,7 @@ create table Review
     createdTime  datetime(6) null,
     memberId     bigint      null,
     modifiedTime datetime(6) null,
-    rating       bigint      null,
+    rating       double precision      null,
     reviewId     bigint auto_increment
         primary key,
     tourItemId   bigint      null,
